requirements.txt
gradio==4.9.0
numpy~=1.26.2
jsonlines~=4.0.0
tiktoken~=0.5.2
tokenizers~=0.15.0
openai==1.3.9
tenacity~=8.2.3
python-dotenv~=1.0.0
Pillow==10.1.0
scikit-learn~=1.3.2
future~=0.18.3
matplotlib==3.8.2